+++ www/Serialization.js
@@ -139,10 +139,20 @@ module.exports = {
     };
   },
 
-  serializeGeotag: function (data, expectedLocation) {
+  serializeGeotagData: function (
+    data,
+    expectedLocation,
+    orderHandle,
+    orderStatus
+  ) {
     return {
       data: data,
       expectedLocation: expectedLocation,
+      orderHandle: {
+        type: "orderHandle",
+        value: orderHandle,
+      },
+      orderStatus: orderStatus,
     };
   },
 };