+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2024-04-23
+
+### Changed
+
+- New `addGeotag` and `addGeotagWithExpectedLocation` methods signature that have `orderHandle` and `orderStatus` parameters. You can use this API when users need to clock in/out of work in your app to honor their work time (see [Clock in/out Tagging](https://hypertrack.com/docs/clock-inout-tracking#add-clock-inout-events-to-a-shift-timeline) guide for more info)
+- Updated HyperTrack SDK iOS to [5.5.2](https://github.com/hypertrack/sdk-ios/releases/tag/5.5.2)
+- Updated HyperTrack SDK Android to [7.5.3](https://github.com/hypertrack/sdk-android/releases/tag/7.5.3)
+
 ## [2.1.2] - 2024-02-28
 
 ### Changed
@@ -339,3 +347,4 @@ We are excited to announce the release of HyperTrack Cordova SDK 2.0.0, a major 
 [2.1.0]: https://github.com/hypertrack/cordova-plugin-hypertrack/releases/tag/2.1.0
 [2.1.1]: https://github.com/hypertrack/cordova-plugin-hypertrack/releases/tag/2.1.1
 [2.1.2]: https://github.com/hypertrack/cordova-plugin-hypertrack/releases/tag/2.1.2
+[2.2.0]: https://github.com/hypertrack/cordova-plugin-hypertrack/releases/tag/2.2.0